Table.js: wrong display text in editor when cell has an error
When a cell editor is opened the text and error of the cell needs
to be copied to the editor so the user is able to adjust the wrong text.

The equals check in setCellValue needs to be removed to ensure the text
is always updated (see CellEditorSpec.js for details). The ValueField.js
acts the same way (see ValueField.setValue).

286460