Table.js: click on a cell does not always open the editor
If a cell editor is open and the user clicks on another cell a new
editor should open for the clicked cell. This does not work when a cell
of the same row is clicked because that row will be rebuilt during
the mouse down handler of the popup resulting in a missing execution
of the mouse down handler of the table.

-> Manually propagate the event to the clicked table row.

287905