Table.js: don't trigger update row event when text does not change
When a cell is edited, or setCellValue() called explicitly,
 the update row event should only be triggered if the cell changes.

The check must not be done in setCellValue itself otherwise
error status handling won't work correctly, see previous commit
and specs.
Instead the check is now added to setCellText method
which is similar to the java code.
This means the columns now need to format the text even if they
don't display it (e.g. boolean column). Otherwise no update event
would be triggered.

Change-Id: Ie046979f233209868aca8c96008838cd5690ada8
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/175905
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>
Tested-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>