Fix for FileChooserField#clear value handling
Before this bug fix the field showed an error status, when the field
had an existing file (in Scout Classic only the displayText has been
set) and the user clicked the clear icon. This happened because
_parseValue was not implemented, thus ValueField#_parseValue was used
which returns a string. Because of that validateMaximumUploadSize
failed later, which expects null or a File object as value.

Since the displayText of the browser input file cannot be changed by
typing, the new implementation returns either null or the file object
from the fileInput.

288702