Remove dependency to org.eclipse.scout.json from the platform
In a previous change, a dependency to
org.eclipse.scout.rt:org.eclipse.scout.json was added to the
org.eclipse.scout.rt:org.eclipse.scout.rt.platform module. This however
lead to org.eclipse.scout.json being added as a dependency to every
project depending on the Scout platform which is not desirable.

This change introduces a new abstraction in the form of the
IJsonPropertyReader interface for JSON string deserialization when
decoding environment variable values as config property maps.

In addition, an implementation for that interface using Scouts
IDataObjectMapper is also provided. In order to make use of this
implementation a dependency to org.eclipse.scout.rt.dataobject and a
dependency providing an implementation for IDataObjectMapper (e.g.
org.eclipse.scout.rt.jackson) is required the application aggregator
module.

Change-Id: Idf3e5a98ff234cae746818aeb85a030fa8aece59
Signed-off-by: Michael Iseli <michael.iseli@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/176819
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>