Generate texts.json and locales.json for plain Scout JS apps
Normally, these files are generated and served by the ui server.
If the Scout JS app does not use a Scout backend it needs
to import the scout texts and locale information somehow.

Also ensure that chart module creates proper artifacts for this use
case, too.

The umd style is removed because it breaks this use case.
When eclipse-scout-core.js is loaded webpack assigns the properties
of the module to the window object. Because the scout module itself
has a variable called scout, window.scout is overwritten by that
variable. Hence, window.scout.App etc. is not available anymore.

The changes in DesktopNotification.less are necessary otherwise
the less build includes empty keyframes which will break the animation.

289432

Change-Id: Iedbf7b4b7efc2068364882cbcf52890211cd85b4
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/176776
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>