Put JavaScript build output in single folder for non-Maven modules
For modules not used as Maven modules the JavaScript output should be
always in the 'dist' directory directly instead distributing it into
'dev', 'prod' and 'res' sub directories.
This ensures that resources requested by relative path (e.g. in css) can
be resolved, as the resources are placed in the same directory.

By default the js build does no longer clean the output directories.
Instead it must be requested with the 'clean' argument. Like this it
behaves the same way as in Maven where the clean is only executed if
requested.

289432