Add maven dependency to eclipse-scout-core for code completion
Previously, there was a non optional (hence transient) dependency
in org.eclipse.scout.rt.ui.html.
This dependency had to be removed because of the new dependency
from eclipse-scout-core > org.eclipse.scout.rt.ui.html that
would create a cycle.

This is only necessary due to the following IntelliJ Bug:
https://youtrack.jetbrains.com/issue/WEB-45410