Fix InputFieldKeyStrokeContext for contentEditable divs
InputFieldKeyStrokeContext tries to detect if the current element is an
input field. In addition to <input> and <textarea> elements, <div> with
contentEditable=true must be considered as well. Otherwise the key
stroke may leak to outer elements and be swallowed.

Additionally, PageTileGridSelectKeyStroke must only be active when the
navigation is in the foreground, i.e. the outline is not in the
background.

287878

Change-Id: I8c3cca924cb2a3d747282222ce4c35dd61612cca
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/176999
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>