MalwareScanner: add possibility to write to a specific directory
For performance reasons it might be desirable to write files to check
for malware to a location other than the default system temp directory.
A new config property "scout.malwareScanner.path" allows to specify
such a location.

IOUtility.createTempFile() method variants were added that allow to
define the parent folder for temporary files.

287374