SmartColumn.js: update event triggers too early
The requiresSave state of a table field is not correctly updated
if the value of a smart column is edited.
This bug was introduced by the commits trying to prevent flickering
(added about a month ago).

The update event must not be fired before the value and the status
are correct.

288929