Fix creation of import declarations in created classes
With update of codemodel from 1.0.3 to 2.3.3 the caching of class
references was removed. The effect was that if a jaxws PortType had
2 identical return values, no import declaration was written in the
generated EntryPoint class. This led to errors in the Callable part
of the generated class.
The new JCodeModelWrapper class ensures the missing caching of class
references. It uses the extended TypeNameParserEx class to work with
cached references and to improve code generation for <?> generic types.

Signed-off-by: Tobias Schlatter <tobias.schlatter@bsi-software.com>
Change-Id: Id6884d8cb5e98a0d429fd8243a435e3d7608f420
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/177579
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Matthias Villiger <mvi@bsi-software.com>