Data object signature test: add initial contribution
The data object signature test creates a signature file representing the
structure (e.g. attributes with their types) of data objects annotated
with a type version. Any changes in such a data object will result in a
test failure to notify the developer that if such a data object is
persisted, a migration might be necessary.

Change-Id: I84730031b62eeee4da8cbdb7e7c3fbc94f623c11
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/178053
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>