TreeSmartField: fix null key behavior
If a lookup row with key=null is used, no lookup rows are shown at all.
This happens if the lookup row with the null key also has a parentKey
set to null. In that case the row is linked to itself and also every
other row points to that one -> no root nodes are found.
It can even create an infinite loop while expanding all nodes.

Solution is to make sure that lookup rows with parentKey = null will
never be linked to lookup rows with key=null.
LookupRows with key=null have a special behavior: if such rows are
selected, the value is set to null and the display text is cleared.

Also needed to check for null key in LocalLookupCall.java
and LookupCall.js to prevent infinite loop if a lookup row uses
null as key and the user searches for the text of that row by typing.

277035

Change-Id: I0312928de41366c872b13159e7735c7ac46bdb31
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/178443
Tested-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>