Outline: too many nodes visible sometimes when table is filtered
If a row is inserted or updated that won't be visible due to an
active filter, the node linked to that row must stay invisible as well.

Insert issue
  The filter event triggered in insertRows() is only fired if the
  visible rows change which is not the case.
  This fails only in Scout Classic. With Scout JS the OutlineMediator
  mediates the insert event to the outline which correctly applies
  the filters.

Update issue
  The update event triggers row init which is handled by the
  OutlineAdapter. That handler applies the outline filters.
  Since at that time row.filterAccepted is still true (default value),
  the DetailTableTreeFilter will accept the node which is wrong.
  -> The filter is triggered too early, in fact, this code is
  partly obsolete, because Table.updateRows() triggers a filter event
  itself that will be propagated to the Outline by OutlineMediator.js.

Unfortunately, the removed code was was necessary for a special case:
If a column that contains filtered cells is removed, the nodes need
to get visible along with the rows. After the code removal the nodes
stayed invisible. This happened because the TableAdapter prevents
filtering during that structure change and the outline won't be
informed.
Fixed by listening for filterRemoved event in OutlineAdapter. Listening
for filterAdded event should not be necessary because inserting the
rows will trigger a filter event anyway.

269409

Change-Id: If9da5eae2f34b8f6dc56d224730c00bb0e49cb49
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/178621
Tested-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>