Add deserializer for Locale that handles Locale.ROOT correctly
Previously, a special serializer for Locale existed, that output the
language tag of a locale instead of the "toString()" value as the
default jackson implementation does. Deserialization, however, was
left to the default implementation. This worked fine, as the default
Jackson deserializer is also capable of deserializing the language
tag into the correct locale. For the root locale, however, this did
not work, i.e. serializing the language tag of the root locale did not
yield the root locale when using the default deserializer.

This change now adds a deserializer (and key deserializer) for Locale,
that only treats the language tag of the root locale ("und",
LanguageTag.UNDETERMINED) in a special way, by returning Locale.ROOT.

Change-Id: I0e22f8a80bf7aa7f478990c6b9986915b18d6b35
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/179618
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Paolo Bazzi <paolo.bazzi@bsi-software.com>