Improve desktop notification and fix specs
hide() now always calls destroy, even if the notification
is not rendered or invisible. This is important for the
nativeOnly case (actively closing a native only notification
 did not remove it from the desktop because hide() was not called).

destroy() now closes the native notification, too, so the user
doesn't have to close both. Closing the regular notification
when the native one is closed still does not happen because
the regular notification might contain additional information
 (e.g. a link).

nativeOnly now does not close the notification immediately anymore
to maintain the link to the native notification. This allows
closing the native notification programmatically by closing the
desktop notification. To ensure the native notification won't be
opened again after a page reload, the nativeNotificationShown flag
was introduced.

Specs fail on Headless Chrome because Notification.permission
returns denied -> replace it to return granted instead.
Also remove unnecessary code from specs and improve test cases.

Use addEventListener syntax instead of onclose/onclick because
it feels better and we always do it that way.

287787