Allow to reset the file-chooser in Scout classic.
When the java model set the value to null, this had no effect on the
INPUT element in the UI. This change has been applied to the Json-
FileChooserButton before. Additionally the order of property change
events in the FileChooserFieldAdapter.js has been fixed, because it
wouldn't be possible to set the value to null and set a display text
at the same time.

290908