Tooltips: improve user experience
tooltips.js:
Reduce the "long" tooltip delay that can be used for cases where there
are other reasons for hovering over an element and fast tooltips could
be annoying. The previous value was 2 seconds, which is much too long
for the user to notice. They give up much earlier and assume that there
is no tooltip at all. The new value is 1 second, which seems to be a
more reasonable value.

Tooltip.js:
When the tooltip text changes while the tooltip is already showing, the
position might need to be adjusted (longer/shorter text). To ensure that
the computed position is always the same, the current position has to be
reset. Mostly relevant when the tooltip is shown at the right side of
the screen.

Button.js/Switch.js:
Like with menus (Menu.js/Action.js), any scheduled tooltip should be
cancelled when the user clicks on the element.