Desktop: update cached navigation width when restoring from breadcrumb
When enlarging the navigation width from the breadcrumb mode back to the
normal mode, the navigation width is automatically set to the minimal
value. However, the cached width was not updated. When the user logged
out and back in, this cached width would be restored, instead of the
previously visible width. When dragging the navigation with the mouse,
the cached size is always updated. The behavior should be the same when
using the collapse handle.