RadioButton: fix layout when useUiWidth=true
When one of the radio buttons in a radio button group has
useUiWidth=true and gridWeightX=0, the layout is not correct, when the
container size changes. Whenever revalidateLayoutTree() is called, the
size of the radio button is increased, instead of staying constant.

Because the radio button size is controlled by FormFieldLayout, it
requires absolute positioning. The previous value was 'relative',
probably by accident (from initial commit / POC code) and the problem
was not noticed since because it requires a special constellation.

274279