Improve desktop notifications
Allow the configuration of title and iconId for native
notifications, either globally or for a specific notification.

This was necessary because it is not possible to always
use the logoId of the desktop. If the desktop uses an
SVG image, the Chrome notification would not show it. Or
maybe the desktop uses an image that is not appropriate
for notifications.

287787