Remove obsolete code
This change removes the last fragments of the so-called "wizard status".
In the Swing/SWT/RAP era, a special HTML file was used to render the
step list and current state of a wizard. With the new HTML UI (version
5.2), this task is now handled by the AbstractWizardProgressField. The
AbstractWizardStatusField and all related classes have not been in use
for several years and can therefore safely be deleted. If you still rely
on these classes, please download them from the Scout source repository
and put them in your own code base.