Table: don't fail if rows are filtered while table is detached
When the user types text in the filter and presses the shortcut
to open the form for the selected row, the filtering may happen
after the table was detached because filtering is started later (using
$.debounce).
If that is the case, the viewport rerendering is blocked, which is fine.
As soon as the table is attached again the table is layouted and the
view port rendered. This will fail because it should actually
re-render it, not just render it.
This only happens if the table itself was detached
(e.g. when it's a detail table), because otherwise _renderOnAttach
would do the rerendering but in this case _postAttach comes first which
calls the layout.

292617

Change-Id: I649120360cddbc61eed624fa44a86150daf54761
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/180229
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>