DesktopNavigation: adjust size of ViewButtonBox in single view mode
In "single view button" mode, the navigation is positioned at the very
top, at the same location as the ViewButtonBox. This looks correct
(ViewButtonBox is transparent), but the outline title cannot be clicked
anymore since the mouse events are blocked by the ViewButtonBox.

To fix this issue, the view button box must be resized to its preferred
size when in "single view button" mode.

273363

Change-Id: Ia185d104020f14bc70923e8520e88295cbf8e3cb
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/180366
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>