Table: fix exception when pressing navigation key on an empty table
This reverts commit 5a9ca395.

The removal of the visibleRows check by the reverted commit leads to an
exception if the user presses a navigation key in an empty table,
because the handle functions don't check for empty table again.

To resolve the original bug (up / down keys don't work in cell editor),
the change pushed to older branches is cherry picked, which is the
change in CellEditorPopup.js from commit d399ce45.

294248