Widget: don't render property changes when removal is pending
If a widget or ancestor widget is being removed by an animation using
animateRemoval, property changes during that time can result in
already rendered exceptions because the removal is prevented but
rendering is not.
Example: Popup (e.g. OrganizeColumnsForm) is being closed and a
selection event occurs (eiter by the user clicking very fast or because
the model sent newly selected rows). When a row is selected the table
needs to render its menus by first removing the old ones and then
rendering the new ones. If the old ones are not removed the rendering
can fail if the same menu should be rendered.

It would probably be possible to just prevent the rendering of a widget
(adding isRemovalPending to render(), which would be consistent
to remove()), but actually every kind of rendering should be prevented
to get a smooth animation. The easiest way to accomplish this is to
convert the rendered property into a real property and doing the
isRemovalPending check there. This way the consumers won't have to
change anything.

Since the removal cannot be aborted it doesn't matter that the model
is not reflected properly during removal.