Widget: remove animation aborts if a child animation finishes
Use case:
A context menu contains a menu that opens a dialog. Activating that
menu draws a glass pane into the context menu because the dialog is
modal. The glass pane has an animation with duration 0 which triggers
the animationend event. The listener in Widget._removeAnimated now
removes the context menu even though only the glass pane animation
ended.
-> Add a check to make sure remove is only executed if the target
is the widget.

Didn't add the check to animationEnd directly since bubbling is the
normal behavior, even though it's probably unexpected here.