jquery-scout: fix oneAnimationEnd() for nested animations
Because "animationend" events bubble up, an event listener will also be
triggered when the animation on a child element ends. To prevent the
registered "oneAnimationEnd" handler to be notified too early, it is
only called when the event target matches the expected element. If the
event originated from a different element, the listener on the current
element needs to remain registered. Therefore, the listener has to be
registered with on() instead of one().

Change-Id: Ic2cd34662b2abc030270a5540b7fff42530cf63b
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/181131
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>