Optimize key strokes
Table:
The removal of the visibleRows check by the commit 5a9ca395 led to an
exception if the user presses a navigation key in an empty table,
because the handle functions don't check for empty table again.
The NPE was (accidentally) fixed by the cherry pick commit 2a14b858.
Even though there is no exception anymore, it still behaves differently
than the tree: if there is an empty table in a search form, pressing
up moves the selection of the outer table. This does not happen when
doing the same in a tree, because the tree accepts the key stroke and
applies the propagation flags, the table does not accept it due to the
visibleRows check -> Since we do not want to refactor the navigation
keystrokes and check for visible rows in every sub class, we just
apply the propagation flags even if there are no rows.

To make sure the key strokes are shown correctly in the described case,
the fix in KeyStroke.js is necessary. When pressing F1, a virtual key
event is created. If this is done for a table navigation key stroke,
the property "shift" would be undefined and therefore never accepted.
This change should only affect F1, because a real key stroke event
never uses undefined.

StringField:
A similar issue can be found in a string field. If a multiline string
field is contained in a search form, up/down won't move the cursor
in the string field. Instead, the selection of the outer table is moved.

294248