Button: fix toggle state with popup
If a menu-button with sub-menus is selected, the hover and active state
are broken -> Added rules to button class, it already contains a rule
for the submenu icon. This does not affect button fields because they
are not selected when the popup is open.

Also added state for default toggle button even though it is rarely
used. The state was already used for a combo menu which makes it
consistent now.