BrushUp: fix menubar flickering
On ScoutJS apps, the menubar flickers when a table row is selected
the first time. This happens because when selecting a row, the menubar
is rebuilt and the first/last markers are lost. The layout will add
them again but since the validation comes later, the invalid state
is visible briefly. To reduce the flickering, the markers are now also
set after the rebuilding. This is a new effect because the first item
now has a left margin. On older releases the menubar had the margin.
If this is not sufficient we need to consider an immediate layout
validation.