Fix keystrokes when clicking inside IFrame, HtmlField or LabelField
When the focus is set inside an IFrame, key events have to be propagated
to the parent document to make the KeyStrokes work as expected.

When the user clicks inside a HtmlField or LabelField (e.g. to select
text), the field should receive the (invisible) focus. Otherwise, the
focus would be set to the desktop, and form key strokes would no longer
work. To prevent the invisible focus being set to these fields by
accident (e.g. when they are the first focusable widget in a form), an
additional check for "tabbable" was added to Widget#isFocusable.

The "selectable" property was added to both HtmlField and LabelField.

280491

Change-Id: I683576bf86ce4393d5567ebda088fda16d990a7d
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/181795
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>