Data objects: add support for DoSet and DoCollection
Currently a data object supports DoValue and DoList nodes. To represent
a set or an unordered collection semantically, DoSet and DoCollection
are introduced. The serialized form of a DoSet and DoCollection is
identically to the one of a DoList (JSON array). If a JSON is
deserialized without any type information, a DoList is used within Java.

Via DataObjectHelper#normalize all DoSet and DoCollection nodes can be
sorted deterministically in order to have the same serialized output.

Change-Id: I4f361b30b2987e23aca64d36607ad27ff08a6e70
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/181931
Reviewed-by: Paolo Bazzi <paolo.bazzi@bsi-software.com>
Tested-by: Stephan Merkli <stephan.merkli@bsi-software.com>