Widget: invalidate layout on css class changes
Toggling a css class could change the size of a widget and could
therefore require a layout invalidation.
From java code, it is not possible to invalidate the layout, so it
would be necessary to create a custom widget and do it with JS.
Or we could add the possibility to invalidate the layout to Widget.java.
But actually calling invalidate layout from outside of a widget should
never be necessary, so we decided to do it on every cssClass change.

If it turns out that it has too many side effects
(flickering, performance) on widgets that don't require the
invalidation, we will think of another solution.

292520