TabBox: tab area size may be wrong with sub-labels
The issue can be reproduced when toggling
sub labels on invisible elements, or when reducing
the size of the tab area so that the tab item
with the sub label goes into the ellipsis.

To fix the first case, the has-sub-label class
must not be added if only the invisible item has
a sub label.
To fix the second case, all tabs need to be invalidated when the sub
label changes on one item to clear the cached
preferred size of those items as well.