Added DataObjectHolder for all IDataObject types (e.g. DoLists)
DoEntityHolder is now a sub class of DataObjectHolder with more
specific types.

Added method DataObjectHelper#getTypeName to read TypeName annotation
from a DO entity at runtime.

Added method throwingMerger() to use the 4 argument version of the
Collectors#toMap method, in order to create a specific HashMap type,
like LinkedHashMap. Instead of declaring a specific lambda each time,
we can now use the throwingMerger function, which is default in the
JRE too, but declared as private, so it cannot be used from outside.

Change-Id: Iad534198aff7da0387bb345d27e25d62e9b0fa02
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/182154
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Andre Wegmueller <awe@bsiag.com>