Data objects: add support for DO entity contributions
DO entity contributions allow to add payload in form of a specific DO
entity to an existing DO entity. The payload is part of an internal
attribute node _contributions. The contributions are designed in a way
that it is known which contribution are applicable to which DO entities
in order to determine a list of DO entities that can be held within
another DO entity.

Change-Id: I632360c0bececb96e8e7def64e9568d2ed91e6dd
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/182811
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Paolo Bazzi <paolo.bazzi@bsi-software.com>