InputField: don't stop propagation for numpad keys with modifiers
ctrl-1-9 activates the desktop tabs. Using ctrl-4 and ctrl-6 on the
numpad does not work while the focus is in an input field. But other
numbers work (e.g. ctrl-2) which is confusing.

This change also fixes a bug introduced by
9e44d2e8b82ff0288319a0b31099fdaa4441189f. Pressing a numpad
key without a modifier must always stop propagation, otherwise entering
number 2 or 8 in an input field of a search form won't work anymore,
since the table filter would grab the focus. This broke
because numpad_2 and 8 were unregistered in setMultiline.

256789