Table: don't always wrap text in a span
Unfortunately, the previous commit broke some behaviour.
The text must not be wrapped if there is no text otherwise
nbsp would never be added and the row would shrink if the cells
are empty. Also, it should not be added if there is custom html because
it is unnecessary and it could break existing css rules.