TileGrid: ensure layout is valid after inserting a hidden tile
If a tile is inserted but not visible (e.g. due to an active filter),
it gets the animate-insert class which prevents the layout validation
-> don't animate invisible tiles

Also, if all tiles are made visible/invisible by a filter, the scrollbar
won't be updated correctly. It works well, if only some tiles are
affected because in that case the TileGridLayout does the scrollbar
update after the move animation => Update scrollbar after the
filter-animation.
Since this is a fast process it is not necessary to do it only once
for all tiles, as it is done when tiles are deleted.

274512

Change-Id: I25ac36fae57eaaffda2b299f5646e83ab353db26
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/182317
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Fritz Schinkel <fritz.schinkel@bsi-software.com>