Add maxRowCount variable to LookupCall.js
Allow lookup calls to use a maxRowCount which will be synchronised with
the SmartField.js browseMaxRowCount.
The LookupCall.js itself does not implement any limiting functions but
extensions may use this property to limit the fetched Lookup rows while
the SmartField.js will still be able to detect if the returned row count
exceeds browseMaxRowCount.

301055

Change-Id: Ic332efafb3c8a122517d4ce7cdf467265b80346f
Signed-off-by: Cyrill Wyss <cyrill.wyss@bsi-software.com>
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/183898
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Andre Wegmueller <awe@bsiag.com>