Added equals function to LookupRow.js
Without the equals function, a property change event for a LookupRow
does not work correctly in SmartField.js, because the selected
lookupRow from the SmartField is another instance than the lookupRow
from the ProposalChooser, even when all properties are the same.

With this change, these two instances will be treated as equals, if
all properties are equals. We use Object.keys because there are
LookupRow instances with additional custom properties.

Change-Id: Id0a6480a8705e4fdb91f0ecf88a4356691e75911
Reviewed-on: https://git.eclipse.org/r/c/scout/org.eclipse.scout.rt/+/183926
Tested-by: Scout Bot <scout-bot@eclipse.org>
Reviewed-by: Andre Wegmueller <awe@bsiag.com>