Add BinaryRefResourceLoader
Add a loader for binaryRefs that resolves the corresponding resource via
the IBinaryRefHandlers. For this purpose any IClientSession of the user
is used so the resources to be loaded must not be dependent on the
session. Resources for which no caching is possible in the browser are
cached on the HttpSession.