Do nothing when _acceptInput is called on SmartField while removing
This issue was noticed because a propertyChange:lookupRow event handler
has been attached to a SmartField. This handler called the insertField()
function on the same group-box which contained the SmartField. This
causes the group-box to remove/render the SmartField. When remove() has
been called, the SmartField couldn't deal with that unexpected state,
which caused wrong property change events for the lookupRow property.

300953