Tile: fix async loading
For example, if a Tile dynamically adds FormFields depending on the
loaded data, the Form of these FormFields is always IForm.CURRENT.get()
evaluated in the current RunContext. If this RunContext is not
explicitly created with the Form in which the Tile is located, then
IForm.CURRENT.get() is not necessarily the Form of the Tile but possibly
a different one. This happens for example when the data is already
loaded while the Form is still being created and has not yet been set.
In this case, operations like addField() fail because the Form of the
Tile and the dynamically created FormFields do not match.
Therefore, the data must always be loaded in the context of the Form in
which the Tile is located. Use IForm.CURRENT.get() only if there is no
parent of type IForm.