Use the rowHeight of the GroupBox for a FormFields status
The FormFieldLayout previously used the rowHeight of the HTMLEnvironment
for the height of the status div. However, if the FormField is in a
GroupBox whose rowHeight has been changed and thus no longer equals the
rowHeight of the HTMLEnvironment, the FormField and status no longer fit
together. Therefore use for the height of the status div also the
rowHeight of the GroupBox if available.