BrushUp: add scroll shadow
The scroll shadow visualizes whether there is scrollable content.
It is installed automatically when scrollbars are installed.
The shadow is a div that is placed after the scrollable.
The div has 4 box shadows, one for each side. Since it is placed
after the scrollable, it has to track the style, visibility and position
of the scrollable and adjust itself accordingly. The size and position
are mainly updated by the already existing scrollbars.update calls by
the layouts. Additionally, dom mutation and intersection observers are
used to track dom movements and visibility changes.

Some alternative solutions that did not work well:
- Using a box shadow inside the scrollable container: The box shadow
  lays under the elements, so if they have a background color the shadow
  won't be visible completely.
- Using position sticky: only works with relative positioning and does
  not work well for horizontal scrolling
- Using position absolute and update position while scrolling as done
  for the scrollbars itself: does not work well with native scrolling
  because scroll events are not fired immediately -> shadow jumps
  around.

292522