Button: improve focus style of link buttons
The focus border is now quite big. If it is the last element
in a scrollable box the bottom focus border will be cut.
Even though it has a big height, it has a small width. This looks
unbalanced.

Add separate :after element that can have a custom size.
Also make focus border for borderless elements a little smaller (use
same inset focus as a menu item).