Use queueMicrotask to validate layout
Microtasks run at the end of the current task but before the browser
does any rendering and painting. This is the perfect time to do our
layouting and makes many validateLayoutTree calls obsolete that
were only needed to prevent flickering. The flickering happened because
the browser is allowed to do the painting between 2 regular tasks, in
our case between the rendering and layouting phase.