Improve memory usage of LookupRow
The text should not be stored as char[] internally. This always creates
a copy of the string. Instead, directly store and return the string
provided. This allows that references are passed instead of copies.
Furthermore, it allows the JRE to apply string specific optimizations
(like deduplication, interning or storing it as byte[] instead of
char[]).

303775