Table: remove 1px top gap on tables without a header
The new selection style draws the selection border over the border of
the previous row, the height of the selection is therefore 1px bigger
than a regular row. Currently, there is a 1px top-padding in the table
data for the selection border of the first row. This creates a gap if
there is no table header which is especially visible if the row has a
background color or if it is editable (triangles are not on top).
Unfortunately the selection cannot overlap the table, so we need to
handle the first row differently.

The solution is to not move the selection up if it is the first row
on a table without a header. To not change the size of the selection
in that case, the selection now overlaps the bottom row instead of the
top one.

This is not necessary for tables without row borders (list box) or if
the first row has a top border (smart field).

Because the cell editor has to be displayed exactly inside the selection
and the text has to be aligned exactly with the text of the table cell,
the editor needed some adjustments.