RestLookupCall: fix restriction precedence
When creating clones of the template RestLookupCall, values predefined
in the model of the lookup call ('restriction' property) should take
precedence over values that are automatically added to
this._restriction in the respective cloneForX method. This behavior was
not well defined before, which could have lead to unexpected behavior.

One example is the 'maxRowCount' property which was always taken from
the lookup call, even when a different value was predefined in the
restriction object.

Documentation for these precedence rules was added to RestLookupCall.js.
Additionally, a Jasmine test was added asserting these rules.

306914