FilterSupport: Add new JS FilterField
Add FilterSupport that can be used to add a JS FilterField for ListBox,
Table, TableField, TileGrid, TileAccordion, Tree, TreeBox and TreeField.
The FilterSupport is also used to handle (add, remove, set) all the
filters for the widget. All functions (add, remove, set) apply
the filters after modifying the filter list by default.
FilterSupport and LoadingSupport are now both a WidgetSupport.
Change KeyStrokes that explicitly use the TableFooter-FilterField to use
the FilterField set on the table. Table- amd TreeNavigationKeyStrokes
focus the widget.
Remove createKey() from TableFilter and move it to TableUserFilter.