Mobile: exception when detail content contains an iframe
If the detail form of a page contains an iframe, Safari
will throw an exception when that page is clicked.
Reason: the microtask queue is processed when the iframe
is inserted, but only in Safari. This is probably a bug since
the current JS code is still being executed.

The exception happens because the validation calls the OutlineLayout
which tries to render the viewport while the viewport and the selected
page including the detail form is still being
rendered -> Already rendered exception.

The current solution is a workaround. Actually, we should try to prevent
layout validations during rendering in general.

309155