Fix 'Resource Busy' error when ui server & watcher are running
Sometimes when the ui server is running and the webpack watcher tries to
update the assets it fails with a 'Resource Busy' error.
This is because the Java process holds a file handle on the web assets.
Basically it reads the content from the disk and caches it in memory but
UrlConnection.getContentLength and UrlConnection.getLastModified in the
background create InputStreams to the files to return this information.
These streams are not closed immediately but somewhere later. If the
watcher tries to write these files it ends with an error.

Furthermore, various JavaScript dependencies have been updated to
include some bugfixes. Most important are bugfixes in the webpack
watcher which are necessary to prevent the 'resource busy' error as
well.

To make the development easier the cache for the file-index (asset list)
is now only available in production mode. This helps that no server re-
start is required if the assets change while the ui server is running.