ViewButton: fix background color when navigation is hidden
If navigation is hidden and the bench
has a menu bar (e.g. if a table is displayed),
the view menu tab will be transparent because the menu bar is
transparent as well.
Since the menu bars are not colorized anymore, it is ok to
just delete the rule. Projects customizing the main menubar
need to add the rule by themselves (and also need to
colorize the edge).

Also fix the edge animation. If the navigation is hidden and
the view menu selected while a dimmed background was active, the
animation was broken -> only change color without redefining
the position.

309392