ViewTabs: display error when views are overflown
If the first view is in the overflow menu, the selected view tab
does not look good.
Reason: the classes 'first' and 'first-tab-selected' are not correctly
updated, because the invisible tabs have not been considered.

Also improve visualization of view tabs inside bench.
The right edge overlapped the splitter on the right. Also, there was a
gap on the last tab if view tabs were as width as the area.

306344